+++ c++/include/computer/Bus.hpp
@@ -22,12 +22,16 @@ namespace computer
         Bus();
         ~Bus();
 
+        void reset();
+        void setPwr(bool pwr);
+
         void run();
         uint8_t get(uint16_t adr);
         void set(uint16_t adr, uint8_t data);
 
         void addDevice(std::shared_ptr<computer::Device> device, uint16_t startAdr, uint16_t endAdr);
         void removeDevice(std::shared_ptr<computer::Device> device);
+        uint32_t getDeviceAdr(std::shared_ptr<computer::Device> device);
     };
 } // namespace computer