+++ c++/include/dynarecs/testDynarecs.hpp
@@ -19,4 +19,6 @@ void testTranslater2();
 
 void testDeviceThread(std::string filePath, uint32_t hz);
 
+void testGraphicDynarec(std::string filePath, bool debug);
+
 #endif // TESTDYNARECS_HPP