+++ c++/include/graphic/BaseWindow.hpp
@@ -1,0 +1,20 @@
+#ifndef BASEWINDOW_HPP
+#define BASEWINDOW_HPP
+
+#include "graphic/Window.hpp"
+
+namespace graphic
+{
+    class BaseWindow : public Window
+    {
+    protected:
+        void start();
+        void stop();
+        void loop();
+    public:
+        BaseWindow();
+        ~BaseWindow();
+    };
+} // namespace graphic
+
+#endif // BASEWINDOW_HPP