+++ c++/src/computer/Bus.cpp
@@ -1,7 +1,10 @@
 #include "computer/Bus.hpp"
 
+#include <iostream>
+
 computer::Bus::Bus()
 {
+    type = DEVICE_TYPE::BUS;
 }
 
 computer::Bus::~Bus()
@@ -10,6 +13,24 @@ computer::Bus::~Bus()
 
 void computer::Bus::run()
 {
+}
+
+void computer::Bus::reset()
+{
+    for (unsigned int i = 0; i < devices.size(); i++)
+    {
+        devices[i]->reset();
+    }
+}
+
+void computer::Bus::setPwr(bool pwr)
+{
+    this->pwr = pwr;
+    for (unsigned int i = 0; i < devices.size(); i++)
+    {
+        devices[i]->setPwr(pwr);
+        std::cout << "pwr["+std::to_string(i)+"]=" << devices[i]->getPwr() << "\n";
+    }
 }
 
 uint8_t computer::Bus::get(uint16_t adr)
@@ -21,6 +42,7 @@ uint8_t computer::Bus::get(uint16_t adr)
             return devices.at(i).get()->get(adr - startAdrs.at(i));
         }
     }
+    std::cout << "no device found\n";
     return 0;
 }
 
@@ -53,4 +75,19 @@ void computer::Bus::removeDevice(std::shared_ptr<computer::Device> device)
             endAdrs.erase(endAdrs.begin() + i);
         }
     }
+}
+
+uint32_t computer::Bus::getDeviceAdr(std::shared_ptr<computer::Device> device)
+{
+    uint32_t res = 0;
+    for (unsigned int i = 0; i < devices.size(); i++)
+    {
+        if (devices.at(i) == device)
+        {
+            res = (startAdrs[i] << 16);
+            res += endAdrs[i];
+            return res;
+        }
+    }
+    return res;
 }