+++ c++/src/computer/CPU.cpp
@@ -4,8 +4,34 @@
 
 #include <vector>
 
+#ifndef _WIN32
+#include <thread>
+#else
+#include "mingw.thread.h"
+#endif
+
 computer::CPU::CPU(std::shared_ptr<computer::Bus> bus) : Device()
 {
+    threadWanted = false;
+    type = DEVICE_TYPE::CPU;
+    this->bus = bus;
+    cycle = 0;
+    pc = 0;
+    hz = 0;
+    for (unsigned int i = 0; i < 16; i++)
+    {
+        reg[i] = 0;
+    }
+    for (unsigned int i = 0; i < 1024 * 64; i++)
+    {
+        rawBus[i] = 0;
+    }
+}
+
+computer::CPU::CPU(std::shared_ptr<computer::Bus> bus, bool threadWanted) : Device()
+{
+    this->threadWanted = threadWanted;
+    type = DEVICE_TYPE::CPU;
     this->bus = bus;
     cycle = 0;
     pc = 0;
@@ -27,7 +53,36 @@ computer::CPU::~CPU()
 void computer::CPU::run()
 {
     dynarec::Translater t(shared_from_this(), false);
-    t.run(pc);
+    if (!threadWanted)
+    {
+        t.run(pc);
+    }
+    else
+    {
+        bool start = false;
+        while (running)
+        {
+            if (pwr)
+            {
+                if (!start)
+                {
+                    start = true;
+                    resetReg();
+                    t.initStep(pc);
+                }
+                t.runStep();
+            }
+            else
+            {
+                if(start)
+                {
+                    start = false;
+                }
+                std::chrono::nanoseconds timeWait(1000000);
+                std::this_thread::sleep_for(timeWait);
+            }
+        }
+    }
 }
 
 uint8_t computer::CPU::get(uint16_t adr)
@@ -42,6 +97,15 @@ void computer::CPU::set(uint16_t adr, uint8_t data)
 
 void computer::CPU::reset()
 {
+    resetReg();
+    if(bus != nullptr)
+    {
+        bus->reset();
+    }
+}
+
+void computer::CPU::resetReg()
+{
     cycle = 0;
     pc = 0;
     for (unsigned int i = 0; i < 16; i++)
@@ -49,6 +113,15 @@ void computer::CPU::reset()
         reg[i] = 0;
     }
     reg[R] = rand();
+}
+
+void computer::CPU::setPwr(bool pwr)
+{
+    this->pwr = pwr;
+    if(bus != nullptr)
+    {
+        bus->setPwr(pwr);
+    }
 }
 
 uint8_t computer::CPU::getBusData(uint16_t adr)