+++ c++/src/computer/Device.cpp
@@ -2,8 +2,26 @@
 
 computer::Device::Device()
 {
+    running = false;
+    pwr = false;
+    type = DEVICE_TYPE::NOVAL;
 }
 
 computer::Device::~Device()
 {
+}
+
+DEVICE_TYPE computer::Device::getType()
+{
+    return type;
+}
+
+void computer::Device::setPwr(bool pwr)
+{
+    this->pwr = pwr;
+}
+
+bool computer::Device::getPwr()
+{
+    return pwr;
 }